refactor user creat view, send emails on sign up/order create
