--- djangoRestReactShop/constants.py
@@ -1,1 +1,0 @@
-# TODO add constants
